Use Left Option (AKA lalt) as modifier
This is the same position as the Super key I'm used to, and it doesn't seem to
cause any conflicts that I can see. To clarify, my keyboard is set to:

 - Caps Lock: Ctrl
 - Fn: Fn
 - Ctrl: Cmd
 - Left Option: Left Option, 'mod' key for Yabai (via skhd)
 - Left Command: Option (AKA Alt)
 - Right Command: Option (AKA Alt)
 - Right Option: Option (AKA Alt)

Only the left Option is picked up by this skhd mod, which is exactly what I
want. If I want an Option (AKA Alt) key then I'll use Left Command, which is
better placed for me. If I want Command I'll use Ctrl, which is better placed
for me. If I want Ctrl I'll use Caps Lock, which is better placed for me.