Fix spaces when switching, if they look wrong
If our spaces aren't set up right (correct number, labelled appropriately) it
can be confusing when we try to switch between them and it goes wrong. The mod-r
binding should fix this, but it would be nice if it were more automatic.

This change adds an extra step before switching between spaces: see if all of
the space labels match what we expect ("l1", "l2", ..., "l9"). If so, we carry
on as normal. If not, that means we either have the wrong number of spaces or
they've got the wrong (possibly empty) labels. Either way, we run the
fixUpSpaces script (the same thing we've bound to mod-r) before we try
switching.

This makes the switching command more complicated, and gives it modification
side-effects; but it's idempotent, so should only be noticable the first time
it's invoked