Try a few times when switching space
Switching space seems to go wrong every now and then; mostly when moving a space
across from another display.

I've added a couple of re-try loops: one for sending the requested space to the
current display, one for focusing the requested space.

This seems to at least be more robust, even if not perfect.