Remove redundant work
Don't always say we're going to label spaces, since we'll skip any that are
already labelled. Instead, only give out a message when we're actually applying
a label.

Secondly, reduce the number of calls to 'yabai -m query --spaces'. We're doing
this *a lot*, when there won't be any changes in between some of these calls.

Instead, store the result in a SPACES variable; then re-use and re-generate it
as needed.