Do a whole bunch more rejigging when trying to focus a space
Trying to switch to a space can fail, for reasons I still haven't figured
out. It might actually be causing yabai to crash (since it seems to keep
restarting).

This commit adds a bunch of extra checks and actions to try and get the space
moved to the correct display and then focused. We retry 10 times, and after a
certain number of retries we will start switching space, switching the display
back and forth, etc. to see if that makes a difference.