Include Haskell stuff, but not stylish-haskell
On the one hand it's a bunch of extra stuff in our environment, but on the other
it minimises divergence from nixos.local

The major headache is bootstrapping GHC for iohk's haskell.nix, but we need to
do that for tools that are useful outside Haskell too, notably nix-diff. If
we're going to pay that price in any case, we might as well get hlint, etc. too.