Add Karabiner Elements config
I'm switching to this so I can get Space Bar to act as Control when held. This
is working :) The current config is:

 - Left Control becomes Left Command
 - Left Command becomes Left Option (Alt)
 - Caps Lock becomes Left Control
 - Space Bar becomes Left Control when held, unless Caps Lock is held

We want CapsLock+SpaceBar to send a Control+SpaceBar event, so that Emacs will
set the mark. This seems to require setting a variable when CapsLock (our
Control key) is being held down, and having our Space Bar rules be ignored if
that variable is set.

The only problem at the moment is that Karabiner is sending the same signals for
both Left Option and Left Command (since they're both set to send Left Option,
that's fair enough); so skhd is being invoked for both.

I think the solution is to stop using skhd and use Karabiner for our keyboard
shortcuts instead.