Add castCommand to reduce the need for redundant pattern-matching
To handle a 'Command m a' we only need to handle 'Command m Bool', since every
constructor of 'Command m a' is specialised to 'Bool'. However, if we stick some
'Bool'-specific things in our handler, GHC will complain that it's not generic
enough.

One way to show GHC that 'Command m Bool' is all we need is to pattern-match,
but that can require duplicating branches. For example:

    foo :: Command m a -> a
    foo = True

The above won't type-check, since GHC can't infer that 'a' will always be 'Bool'

    bar :: Command m a -> a
    bar FocusWindow _ -> True
    bar FocusSpace  _ -> True
    ...

The above will type-check, since each return value is given in a context where
the return type has unified with 'Bool'. Yet this isn't ideal, since (a) we have
to write a whole load of pattern-match boilerplate, which will need to be kept
in sync with changes to the datatype (e.g. if we add a new 'Command'); and (b)
the return value that we're duplicating might be a lot more complicated than
'True', which is another case of DRY.

To avoid this I've added a function 'castCommand :: Command m a -> Bool -> a'
which returns the given 'Bool', but cast to type 'a'. This is safe, since 'a'
will always be 'Bool' due to the constructors; it's just that GHC doesn't notice
this. This also restricts the pattern-match boilerplate to a single place
(except for pattern-matching that we *want* to do, of course). It also prevents
duplicating the code for return values, since we just need to give it once as an
argument to the 'castCommand' function.