Improve Yabai tests with more specific test state
We were using the Yabai query results as our test state, but that has lots of
redundancy (e.g. SpaceInfo says which Display it's on, whilst DisplayInfo lists
it's Spaces, and so on). This made life hard as we tried to maintain consistency
between all of these things.

We now use a much simpler approach: we use new types for our test state, which
store only what is necessary. DisplayInfo, SpaceInfo and WindowInfo are derived
from these as needed. In particular, instead of [a] we use the type 'NEZipper a'
which is a non-empty Zipper. This ensures, by construction, that:

 - It's non-empty, e.g. there's at least one Display and each Display has at
   least one Space.
 - There is always one "focused" element. We use this for:
  - The Display which (contains the Space which) is focused
  - The Space which is visible on a Display
  - The Window which is focused on a Space

The only invariant which we don't guarantee by construction is that Windows
should all have distinct IDs. I couldn't figure out a way to encode this, which
would:

 - Preserve Window IDs when moving them between Spaces, e.g. we can't enforce
   that Window IDs on later Displays/Spaces are higher than earlier ones, since
   that would forbid moving a Window on to an earlier Display/Space
 - Not be really complicated to calculate/enforce. For example we might use some
   context-dependent counter, which skips numbers that have appeared earlier
   (e.g. [0, 0, 2, 0] would be IDs [0, 1, 4, 2]), but this would require
   adjusting a whole bunch of Windows when we move one of them.

Since Windows are effectively the 'leaves' of our state, and none of the Yabai
commands change them once created, it's not too big of a deal to leave this
unenforced.

When generating a state we use a post-processing pass to renumber duplicate
Windows. We don't adjust Window IDs during shrinking, so duplicates can't arise
there.