Populate an "Applications" folder on activation
nix-darwin already puts symlinks to installed applications into ~/Applications,
but that is a symlink to a folder, not a folder itself. This means:

 - If we add one of those application symlinks to the dock, it won't get updated
   when we 'darwin-rebuild switch' to a new version
 - If we add the Applications symlink to the dock, it won't get updated either
 - A symlink to a folder will cause Finder to open, rather than a "stack" (popup
   of contents above the dock)

For these reasons I've created a new folder ~/.local/share/Applications (chosen
mostly arbitrarily), which has symlinks to the symlinks in ~/Applications. An
activation script (run at boot and at the end of 'darwin-rebuild switch') clears
out its contents and re-adds the symlinks.

Adding this to the dock ensures that:

 - We can open applications quickly (via the "stack" popup); although it's not
   one click
 - The applications get updated when we switch to a new build of the system
 - Removing applications from the system configuration will remove them from the
   ("stack" in the) dock