Use upstream yabai module
Yabai is now in upstream nixpkgs (unstable), and cmacrae's yabai service is now
in upstream darwin-nix (with a slightly different config).

Note that if trying to upgrade, there might be complaints about the
yabai.service either not being defined, or having duplicate definitions. This is
probably due to old entries in the shell's NIX_PATH. Try commenting out the
Yabai stuff, rebuilding, then refreshing the shell variables (e.g. starting a
new shell, or copying them across from a new shell, or whatever). The service
can then be re-enabled.