Call 'run-healthcheck' from within 'healthcheck'
It's tricky to do nested stdin-reading 'while' loops in bash, since they all try
to read from the same stdin stream. In this case it's easier to pull out the
healthcheck finding and testing into a separate 'run-healthcheck' process, and
keep the 'healthcheck' process for doing the stack/api finding.