Patch aws-sam-cli dependencies to match nixpkgs versions
This package was refusing to build after updating nixpkgs. We could package up
the specific versions of dependencies it needed, but that's quite a rabbit hole
to go down (packaging their dependencies, and so on). The existing package
already patches some of the version numbers in requirements/base.txt, so we just
add a few more.